map circles modularised and sizing fixed
